move media form component and update llm model
